refactor: organize py printer src files
