+++ Solverz/code_printer/__init__.py
@@ -1,0 +1,2 @@
+from Solverz.code_printer.python.inline.inline_printer import made_numerical
+from Solverz.code_printer.make_module import module_printer
